button layout

+++ src/org/discobots/recyclerush/commands/SetLiftCommand.java
@@ -8,8 +8,8 @@ import edu.wpi.first.wpilibj.command.Command;
  *
  */
 public class SetLiftCommand extends Command {
-	float liftSPD;
-    public SetLiftCommand(float liftSPD) {
+	double liftSPD;
+    public SetLiftCommand(double liftSPD) {
         // Use requires() here to declare subsystem dependencies
         // eg. requires(chassis);
     	requires(Robot.liftSub);
